requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
datasets==3.2.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.9.0
fugashi==1.4.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
iniconfig==2.0.0
ipadic==1.0.0
Jinja2==3.1.5
jiter==0.8.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
openai==1.59.6
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
plac==1.4.3
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==18.1.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rhoknp==1.3.0
rich==13.9.4
safetensors==0.5.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
SudachiDict-core==20241021
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.3.0
wasabi==0.10.1
Werkzeug==3.1.3
wrapt==1.17.1
xxhash==3.5.0
yarl==1.18.3
